Correction affichage image. Ajouts bouton supprimer tuto dans la page de...

See merge request deptinfo.cegepgarneau.ca/gwenael!88

+++ src/Web/Pages/Tutoriel/Consultation.cshtml.cs
@@ -1,12 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
+﻿using Gwenael.Domain;
+using Gwenael.Domain.Entities;
+using Gwenael.Web.FctUtils;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using System.Linq;
-using Gwenael.Domain;
-using Gwenael.Domain.Entities;
 using Spk.Common.Helpers.String;
+using System;
+using System.Collections.Generic;
+using System.Data;
+using System.Linq;
 
 
 namespace Gwenael.Web.Pages
@@ -23,34 +24,69 @@ namespace Gwenael.Web.Pages
             public Tutos tutoriel { get; set; }
             public List<RangeeTutos> lstRangeeTuto { get; set; }
             public string id { get; set; }
+            public bool droitAccess { get; set; }
         }
-        public ConsultationModel(GwenaelDbContext pDb) => _db = pDb;
+        public ConsultationModel(GwenaelDbContext pDb)
+        {
+            _db = pDb;
+        }
 
+        public Guid ObtenirIdDuUserSelonEmail(string email)
+        {
+            User user = (User)_db.Users.Where(u => u.UserName == email).First();
+            return user.Id;
+        }
         public IActionResult OnGet()
         {
+            if (User.Identity.IsAuthenticated)
+            {
+                Guid idConnectedUser = ObtenirIdDuUserSelonEmail(User.Identity.Name);
+                if (Permission.EstGestionnaireDeContenu(idConnectedUser, _db))
+                    Input.droitAccess = true;
+                else Input.droitAccess = false;
+            } else Input.droitAccess = false;
+
             Input = new InputModel();
 
             if (Request.Query.Count >= 1)
             {
                 Input.id = Request.Query["id"];
-                if (IdEstValide()) return Page();
+                if (IdEstValide())
+                {
+                    return Page();
+                }
             }
 
             return Redirect("/tutoriel");
         }
 
-        public IActionResult OnPost() => Page();
+        public IActionResult OnPost()
+        {
+            return Page();
+        }
 
-        public IActionResult OnPostRedirectHomeTuto() => RedirectToPage("Index");
+        public IActionResult OnPostRedirectHomeTuto()
+        {
+            return RedirectToPage("Index");
+        }
 
         public IActionResult OnPostDeleteTuto([FromBody] TutorielIdVal tutoVal)
         {
             try
             {
-                _db.Tutos.Remove(entity: _db.Tutos.Where(t => t.Id == Guid.Parse(tutoVal.tutorielIdVal)).First());
-                _db.SaveChanges();
+                if (User.Identity.IsAuthenticated)
+                {
+                    Guid idConnectedUser = ObtenirIdDuUserSelonEmail(User.Identity.Name);
+                    if (Permission.EstGestionnaireDeContenu(idConnectedUser, _db))
+                    {
+                        _db.Tutos.Remove(entity: _db.Tutos.Where(t => t.Id == Guid.Parse(tutoVal.tutorielIdVal)).First());
+                        _db.SaveChanges();
 
-                return Redirect("/tutoriel?deleteStatus=true");
+                        return Redirect("/tutoriel?deleteStatus=true");
+                    }
+                }
+                return Redirect("/tutoriel");
+
             }
             catch (Exception)
             {
@@ -62,10 +98,18 @@ namespace Gwenael.Web.Pages
         {
             try
             {
-                _db.Tutos.Where(t => t.Id == Guid.Parse(tutoVal.tutorielIdVal)).First().EstPublier = false;
-                _db.SaveChanges();
+                if (User.Identity.IsAuthenticated)
+                {
+                    Guid idConnectedUser = ObtenirIdDuUserSelonEmail(User.Identity.Name);
+                    if (Permission.EstGestionnaireDeContenu(idConnectedUser, _db))
+                    {
+                        _db.Tutos.Where(t => t.Id == Guid.Parse(tutoVal.tutorielIdVal)).First().EstPublier = false;
+                        _db.SaveChanges();
 
-                return Redirect("/tutoriel?unPublishStatus=true");
+                        return Redirect("/tutoriel?unPublishStatus=true");
+                    }
+                }
+                return Redirect("/tutoriel");
             }
             catch (Exception)
             {
@@ -82,10 +126,9 @@ namespace Gwenael.Web.Pages
         {
             if (!Input.id.IsNullOrEmpty())
             {
-                //Input.tutoriel = _db.Tutos.Where(t => t.Id == Guid.Parse(Input.id) && t.EstPublier == true).Any()();
-                if (_db.Tutos.Where(t => t.Id == Guid.Parse(Input.id) && t.EstPublier).Any())
+                if (_db.Tutos.Where(t => t.Id == Guid.Parse(Input.id)).Any())
                 {
-                    Input.tutoriel = _db.Tutos.Where(t => t.Id == Guid.Parse(Input.id) && t.EstPublier).First();
+                    Input.tutoriel = _db.Tutos.Where(t => t.Id == Guid.Parse(Input.id)).First();
                     GetContenue();
                     return true;
                 }
@@ -93,6 +136,9 @@ namespace Gwenael.Web.Pages
             return false;
         }
 
-        private void GetContenue() => Input.lstRangeeTuto = _db.RangeeTutos.Where(r => r.TutorielId == Guid.Parse(Input.id)).ToList();
+        private void GetContenue()
+        {
+            Input.lstRangeeTuto = _db.RangeeTutos.Where(r => r.TutorielId == Guid.Parse(Input.id)).ToList();
+        }
     }
 }
